Reworked Field component, fixed country order
