+++ src/app/component/CategorySort/CategorySort.style.scss
@@ -30,19 +30,19 @@
         }
     }
 
-    .Field {
+    .FieldSelect {
+        &::after {
+            @include before-desktop {
+                left: 55px;
+                bottom: 22px;
+            }
+        }
+
         &-Select {
             @include before-desktop {
                 position: absolute;
                 top: -35px;
                 opacity: 0;
-            }
-        }
-
-        &-SelectWrapper::after {
-            @include before-desktop {
-                left: 55px;
-                bottom: 22px;
             }
         }
