+++ src/app/component/Field/Field.config.js
@@ -20,7 +20,3 @@ export const PASSWORD_TYPE = 'password';
 export const SELECT_TYPE = 'select';
 
 export const ENTER_KEY_CODE = 13;
-export const A_KEY_CODE = 65;
-export const z_KEY_CODE = 122;
-export const Z_KEY_CODE = 90;
-export const a_KEY_CODE = 97;
