+++ src/app/component/FieldInput/FieldInput.component.js
@@ -1,7 +1,18 @@
+/**
+ * ScandiPWA - Progressive Web App for Magento
+ *
+ * Copyright © Scandiweb, Inc. All rights reserved.
+ * See LICENSE for license details.
+ *
+ * @license OSL-3.0 (Open Software License ("OSL") v. 3.0)
+ * @package scandipwa/base-theme
+ * @link https://github.com/scandipwa/base-theme
+ */
+
 import PropTypes from 'prop-types';
 import { PureComponent } from 'react';
 
-class Input extends PureComponent {
+export class FieldInput extends PureComponent {
     static propTypes = {
         formRef: PropTypes.oneOfType([
             PropTypes.func,
@@ -28,4 +39,4 @@ class Input extends PureComponent {
     }
 }
 
-export default Input;
+export default FieldInput;
